requirements.txt
openai
ollama
requests
langchain_community
fastapi
spacy
uvicorn 
pdfplumber 
python-docx 
spacy 
sentence-transformers
faiss-cpu
pypdf
langchain-ollama
langchain-community
langchain_google_genai
pyPDF2
langchain-community 
google-generativeai
python-multipart